refactor(View): connect ScannerView & ImageCropView
